% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCALPEL_helper.R
\name{summary}
\alias{summary}
\alias{summary.scalpelStep0}
\alias{summary.scalpelStep1}
\alias{summary.scalpelStep2}
\alias{summary.scalpelStep3}
\alias{summary.scalpel}
\title{Summarize results from SCALPEL pipeline.}
\usage{
\method{summary}{scalpelStep0}(object, ...)

\method{summary}{scalpelStep1}(object, ...)

\method{summary}{scalpelStep2}(object, ...)

\method{summary}{scalpelStep3}(object, ...)

\method{summary}{scalpel}(object, step, ...)
}
\arguments{
\item{object}{An object returned by one of the SCALPEL functions:
\code{\link{scalpel}}, \code{\link{scalpelStep0}}, \code{\link{scalpelStep1}}, \code{\link{scalpelStep2}}, or \code{\link{scalpelStep3}}.}

\item{...}{Additional arguments to be passed, which are ignored in this function.}

\item{step}{The SCALPEL step (0, 1, 2, or 3) that you wish to summarize. This is only needed if summarizing an object of class \code{scalpel}.}
}
\value{
None
}
\description{
Prints the parameters used and a summary of results for a specified step of SCALPEL.
}
\examples{
\dontrun{
### many of the functions in this package are interconnected so the
### easiest way to learn to use the package is by working through the vignette,
### which is available at ajpete.com/software

#assumes you have run the examples for the "scalpelStep0" and "scalpel" functions
summary(Step0Out)
 #summarize each step
 summary(scalpelOutput, step = 0)
 summary(scalpelOutput, step = 1)
 summary(scalpelOutput, step = 2)
 summary(scalpelOutput, step = 3)
}
}
